--- .rubocop_todo.yml
@@ -1,51 +1,0 @@
-# This configuration was generated by
-# `rubocop --auto-gen-config`
-# on 2019-06-20 19:14:01 -0700 using RuboCop version 0.58.2.
-# The point is for the user to remove these configuration records
-# one by one as the offenses are removed from the code base.
-# Note that changes in the inspected code, or installation of new
-# versions of RuboCop, may require this file to be generated again.
-
-# Offense count: 1
-Airbnb/FactoryAttrReferencesClass:
-  Exclude:
-    - 'spec/factories/topic.rb'
-
-# Offense count: 1
-Airbnb/FactoryClassUseString:
-  Exclude:
-    - 'spec/factories/topic.rb'
-
-# Offense count: 9
-Airbnb/OptArgParameters:
-  Exclude:
-    - 'lib/kafkat/command.rb'
-    - 'lib/kafkat/command/cluster_restart.rb'
-    - 'lib/kafkat/interface/zookeeper.rb'
-    - 'lib/kafkat/utility/formatting.rb'
-
-# Offense count: 4
-Airbnb/RspecDescribeOrContextUnderNamespace:
-  Exclude:
-    - 'spec/lib/kafkat/command/cluster_restart_spec.rb'
-    - 'spec/lib/kafkat/command/drain_spec.rb'
-    - 'spec/lib/kafkat/command/verify_replicas_spec.rb'
-    - 'spec/lib/kafkat/command_spec.rb'
-
-# Offense count: 3
-Airbnb/SimpleModifierConditional:
-  Exclude:
-    - 'lib/kafkat/command.rb'
-    - 'lib/kafkat/command/topic_alter_reassign.rb'
-    - 'lib/kafkat/command/topic_verify.rb'
-
-# Offense count: 2
-RSpec/ScatteredSetup:
-  Exclude:
-    - 'spec/lib/kafkat/command_spec.rb'
-
-# Offense count: 49
-# Configuration parameters: AllowHeredoc, AllowURI, URISchemes, IgnoreCopDirectives, IgnoredPatterns.
-# URISchemes: http, https
-Metrics/LineLength:
-  Max: 139
