+++ lib/kafkat/utility/formatting.rb
@@ -2,7 +2,7 @@
 
 module Kafkat
   module Formatting
-    def justify(field, width = 2)
+    def justify(field, width: 2)
       field = field.to_s
       count = [width - (field.length / 8), 1].max
       field + "\t" * count
